Rework initialization logic
Previously, the API version check and the OAuth token
retrieval/refresh were made at the first request. This was confusing
for some users. This commit moves these actions to the initialization
logic of DracoonClient. Now, the creation of DracoonClient fails, if
e.g. the API version is not supported or the authorization data is
invalid.

See: #SDKJAVA-73